Add selectable example and some todos
