+++ apps/cookbook/src/app/examples/item-example/item-example.component.html
@@ -13,5 +13,5 @@
 <h2>Nested Controls</h2>
 <cookbook-item-example-nested-controls></cookbook-item-example-nested-controls>
 
-<h2>Disclosure</h2>
-<cookbook-item-example-disclosure></cookbook-item-example-disclosure>
+<h2>Selectable & Disclosure</h2>
+<cookbook-item-example-selectable></cookbook-item-example-selectable>
